sim: initial checkin of simulator
This version runs but has severe restrictions, such as only allowing
files in the root for a user. Although untested, it should allow multiple
users.

Many TODOS, including completing functionality, breaking it into
packages, adding encryption.

This serves as much as a test of our new git setup as of the simulator.