Update patch set 1
Patch Set 1: Code-Review+2

This is fine with me as a change from before, but why use horrible base64 with special case handling instead of the much simpler to use hexadecimal string? Space is not an issue.

But this is fine for now

Patch-set: 1
Reviewer: Rob Pike <5005@0759ed4b-3fdc-34dc-866b-f2e708168c03>
Label: Code-Review=+2