Update patch set 1
Patch Set 1:

Problem with calling Sum256 is that I need the whole thing in memory. With this mechanism, I'm moving blocks to disk as fast as they come in and I don't keep more than X MB in memory. This will be necessary for high scalability, I think.

Thanks.

Patch-set: 1
Label: Code-Review=0