client/testclient: start of test client package
This implements and tests basic operation of Get, Put, and MakeDirectory.
There are many TODOs, most of which are questions we need to discuss.

Also including an untested implementation of File. Ignore that for review;
it's only here so the code compiles. I'll test and update that in the next CL,
to keep this one modest.

Change-Id: If23856413bcd67cfddaf29f1972ca81af64b429e