client/testclient: test File implementation
This adds simple tests for the File implementation of the test client.
The implementation itself is also ready for code review now.

Other than the long list of TODOs, and perhaps a few more tests,
this finishes the testclient implementation, absent crypto and absent
the right story about how to stitch together a working system.

Change-Id: Ie78bcdf2581a754aebfd3be5c742c830e7372f54