client/eeclient: initial checkin
This is a draft for how to update upspin API for handling crypto metadata.
client.go provides an outline of how the crypto would work, but is not
a serious attempt at the final Client.   In particular, it leaves out
the Location loop, but more importantly does not provide a great way
to select packing.   It also fails to implement any of the actual crypto
except for AES.

Change-Id: I5b16e34b7d5973e4b6b9be41f298778347ec36e7