pack: add a registry for packing algorithms and a trivial no-op Packing
Proposal for the design.

We may decide to allocate buffers for the returned data from Pack and Unpack
but this proposal requires the caller to provide the buffer.

I worry about what to do with an unrecognized Packing. We may end up
providing some sort of default set of Packings automatically installed.

Change-Id: I344355e107e86eee501ed841fd9caac2052c1cf8
Reviewed-on: https://upspin-review.git.corp.google.com/1134
Reviewed-by: Eduardo Pinheiro <edpin@google.com>