upspin: change Store.Get and store.Put to use just keys, not locations
As discussed yesterday, this is a great cleanup of the model for the
Store and separates considerations of packing completely from the
store, which is now a pure blob get-and-put model.

Update the test store, directory, and client to follow the new API.
These are cleaned up considerably by the change.

I left the gcp stuff alone, since it seems to need some redesign.

Change-Id: I6a458895c96cd7e6ea064f76f23a475093803edf
Reviewed-on: https://upspin-review.git.corp.google.com/1151
Reviewed-by: Eduardo Pinheiro <edpin@google.com>