all: use the access methods properly in {user,directory,store,client}/test*
Remove where possible knowledge of one test component in another.
(One that must remain is that testdir knows to store its data in teststore.)
This gets the pieces much closer to honest separable components.

The directory service still must be updated not to assume the user
data is in the test store.

Change-Id: I67e37ba1f9c46f9ac9b8593434fb2e6266fab80b
Reviewed-on: https://upspin-review.git.corp.google.com/1153
Reviewed-by: Eduardo Pinheiro <edpin@google.com>