client/testclient: remove most TODOs from the file implementation
Mostly this involves dealing with offset overflow, made tricky because
the API (twinned from os.File's) uses int64s for file offsets but we store
the full file in a slice, which is only int-sized. On 32-bit machines that's
a problem.

Also test that we can seek past the end of a file and it is zero-filled.

Change-Id: If2df8c2fb556b1a90a300ba303ae2364c305b605
Reviewed-on: https://upspin-review.git.corp.google.com/1160
Reviewed-by: Eduardo Pinheiro <edpin@google.com>