ee/ee.go: crypto Packers
status:
ee_test.go passes, but the crypto is not fully tested yet.
The key files are now in ~/.ssh.   Run ../keygen/keygen.go.

Rob, I found a bug in what we were discussing yesterday.
How do I pass dst slice header by reference rather than by
value?  That is, I want to update len(dst) in pdBytes but
couldn't figure out how.  I rewrote the code so it works,
but the style is ugly.  I expect to do more of this in the
future so would love to learn the right idiom.

Change-Id: I4264a6c100ec9585724375d04e12d448ed881608