all: define a more realistic upspin.ClientContext and deploy it through the test pieces
This cleans up a lot of the interaction between the components.
I did not touch the GCP code because I believe it needs a little
readjustment, estalishing more information during setup as opposed
to passing it around with the context.

This does not flow the context into the Packer, but that will be easy
to do.

Change-Id: Iecf8b6c87318501041fa35a06ad8bd24dac38eba
Reviewed-on: https://upspin-review.git.corp.google.com/1205
Reviewed-by: Eduardo Pinheiro <edpin@google.com>