Update patch set 1
Patch Set 1:

You can't do ls -l without the DirEntry. You can't even tell if a file is a directory.

I believe the Client interface should be sufficient for almost everything, if not truly everything. Otherwise the user will reach behind the Client to do things and that might make caches invalid and so on. Bad karma.

Patch-set: 1
Label: Code-Review=0