all: enforce constraints on metadata in Packers
This pushes the proposed changes through, and documents them.
Everything passes but the testdir and testclient implementations
use DebugPack only, because fixing them is a fair bit more code
that deserves a separage CL.

Added and applied helpers to the pack package. Fixed all the
crashes found by the checks.

Disabled another case in test/all_test.go because now the checking
finds cheaters.

Renamed testpack to debugpack, which is more accurate.

Changed the type of the argument to Put to be PackData.

Finally, deleted a spurious type for a pack constant in upspin.go.
They should either all be Packing or untyped. I went with untyped
because the constant behave more nicely that way, but it's easy
(if noisy) to change it later.

Change-Id: I7c77c2371bf4bfaab02d342809c92145b7a843a6
Reviewed-on: https://upspin-review.git.corp.google.com/1295
Reviewed-by: Eduardo Pinheiro <edpin@google.com>