Update patch set 3
Patch Set 3:

(5 comments)

All done.

I'll submit this since it only breaks a test that requires a special flag to run. When we (Rob or I or both) fix a hack in testclient, the test will fail and I'll fix that test. It's actually a good thing to have the test pre-broken so we'll know when we've fixed the hack.

Thanks.

Patch-set: 3
Label: Code-Review=0