context: Get rid of LoadContextWithRC.
It isn't necessary if we just change the default to ~/upspin/rc
from ~/.upspin/config.

Also, forget about using user.Current().  Rob was right, $HOME
makes more sense since (1) posix requires it to be set and
(2) it is the right answer for setuid programs.  Not sure what
happens when we hit non Posix OSes but we'll have to change lots
of other assumptions too.

Change-Id: I8f913a436ed0954ff717e206dd87f4e79daaa558
Reviewed-on: https://upspin-review.git.corp.google.com/1433
Reviewed-by: Eduardo Pinheiro <edpin@google.com>
Reviewed-by: Rob Pike <r@google.com>