key/sha256key/sha256.go: fix comment
(just fixing the comment in this CL;  editorial opinion follows:)

I'm sad in a way that sha256key exists;  I wish that packages hash
and crypto/sha256 and built-in []byte handling were so straightforward
to use that we didn't need the layer of indirection.

On the other hand, there is some possibility that security reviewers
will ask us to switch from sha256 to hmac-sha256.  That will be
easier if we've consistently used sha256key everywhere.

Change-Id: Ifd2b5201b83c808b4e21e7ae3f509f614034e3bf