user/gcpuser: no need to make a copy of the key vector
We can JSON decode directly into the destination type (upspin.PublicKey);
there is no need to do it into []byte and then make another copy.

Ironic that the last CL made a copy and this one removes one, but in this
case it's a nonce value that is not recorded locally so can be handed directly
to the caller.

Change-Id: I9d94f1cbb65b2edbe2ed08e836781814b073edb6