upspin: add a PutOptions struct to be used in Directory.Put
This also requires adding a Time type. Many choices available, but went
with signed 64-bit seconds since the Unix epoch, since this is familiar
to programmers and is less likely to confuse than other options that
may be rationally superior.

Added formals to all implementations of Put and put nil arguments on
all calls. All tests. pass.

Nothing done to the metadata yet.

Change-Id: I4c0a277a437447ad607c20db736ef0c343e039c4