all: change Key to Reference and give it a type
The previously named lookup Key field is easy to confuse with the keys we are using
for security, so rename it Reference, which used to be the name of the struct that
contained it.

Give it a type and push the types through so everything builds again.

Change almost all mentions of the old key to "ref" or "reference". A few places
were not done, most importantly in the field used to send a JSON blob to the
GCP server, since the name is semantically important. Still, once we can push
an updated server, it would be nice to rename the field.

All tests pass. Go vet is quiet.

Change-Id: Icb412a6a7e68404aaf79b69eb7e7d2cd20b0a328
Reviewed-on: https://upspin-review.git.corp.google.com/1510
Reviewed-by: Eduardo Pinheiro <edpin@google.com>