pack/ee/ee.go: AES 256
INCOMPATIBLE FILE FORMAT CHANGE

This is the first substantive CL based on security design review.
Although AES 128 is still generally considered good enough for single
files, we have to assume that public clouds will be subject to brute
force attack in parallel across all files with many keys, which some
claim gives an amortized cost more like 2^80 than 2^128.  To forestall
debate and to be consistent with GCP's decision to go to AES-256
everywhere, we also move to it even for P-256.  This means 14
rounds rather than 10 rounds, both implemented in asm, so the
slowdown is acceptable.

Change-Id: I51d80cf1a58434a765068a9f2412dd2d66801c65
Reviewed-on: https://upspin-review.git.corp.google.com/1541
Reviewed-by: Rob Pike <r@google.com>