cmd/dfuse: fix Setattr so that file truncation works.
FUSE uses Setattr to truncate files rather than a flag
on Open. FUSE will present the change immediately to
users of FUSE.  However, it will be propogated to
the Directory and Store only when all writers close
the file.

I also removed a few hacks that existed till GCP added
file lengths to the directory entry and allowed lookups
of the user root.

Change-Id: I3510e27aabeea48aa1177ee4c2e046fd65ad6d6e