cmd/dfuse: make dfuse actually obey some local consistency rules.
In particular, all local opens of a file will see the same data.
The file will propogate to upspin store on any close in which the
file is dirty, i.e., doesn't match what was read out of upspin store.

I had to change the structure of the code a bit to get there.  cache.go
contains all the code to maintain the cache.

To avoid consistency problems, all mutations of a file are serialized
using the mutex in the node struct.

Change-Id: I3510e27aabeea48aa1177ee4c2e046fd65ad6d6e