upspin: change signature of Directory.Put to take only a Location
To avoid passing the data through the Directory to the Store, have
Directory.Put take a Location. The Client thus writes the data to the
store, then tells Directory where it has been stored.

There is now no reason for Directory.Put to return the Location, so
its signature simplifies.

Because the Directory no longer sees the data, it cannot set the
default value of Size to the size of the data. The default is thus 0,
but we expect Client to use PutOptions to set the size anyway.

I have updated everything that is non-GCP.

Change-Id: I2cdbe109e83a9f6f122271c99d1d0d87b5fc0fb0
Reviewed-on: https://upspin-review.git.corp.google.com/1593
Reviewed-by: Eduardo Pinheiro <edpin@google.com>