access: new API, implement groups
As discussed in mail, this is a new API for the access package.
There is now an Access type with a Can method that checks
permissions, including evaluating groups. If it needs a group
file that it has not already seen, Can returns its name; the caller
can than load it with AddGroup and retry the Can call.

Parsed groups are stored in a global cache, since many
access files may share a group's information.

Also add a Right type that expresses the desired rights (Read,
Write, etc.). These are part of the Can API but also let some
code be simpler.

Further testing will be necessary but the basics are working.

I have not updated any code to use this, but will do so in a
separate CL.

Change-Id: I080c1d5f0ebf8049ecdb0cbeb8110039ab78d049