dfuse: lots of independent changes to improve dfuse speed.
directorycache.go represents cached user Directories. It keeps us
from constantly binding to the directory each call which represents
a round trip to the server for GCP. However, so that we don't get
screwed by a stale bind, if a Directory.Lookup fails, we clear the
cache and try again.

We also do a usercache.Install() to make sure we don't end up
talking to the user server every call.  That cache is an LRU
and entries time out.

Change-Id: I09096afdfeb3d9c775474d5936e25b0345846189
Reviewed-on: https://upspin-review.git.corp.google.com/1660
Reviewed-by: Rob Pike <r@google.com>