pack/debug: look up keys when signing
Sign (and verify the signature) using the appropriate key for the owner
of the file. We don't bother to install a signature for each user, so we
must use that of the one user we know for sure is there, the owner.

Previously it always used the keys provided by the Context, which is
wrong if the caller and the owner differ.

Update affected code, and while we're at it tidy up the access tests
in testdir, which uncovered this problem.

Change-Id: I1f1d7d6f4e4564e1ab6720722073f69284b9db58
Reviewed-on: https://upspin-review.git.corp.google.com/1710
Reviewed-by: Eduardo Pinheiro <edpin@google.com>