upspin/upspin.go: dir Put readers,paths; Share
This change reflects our discussed redesign from AddWrap to Share.

I'm not satisfied that the code is as robust as it should be or
that there are enough tests yet, but it's certainly better than
before.  I'd like to get this reviewed and submitted before we
drift apart with all the other changes in flight.

This update cleans up some TODOs and eliminates spurious
NewFactotum calls.

Change-Id: I3772e897896c808150a5df2a30fa3710ce2f5843
Reviewed-on: https://upspin-review.git.corp.google.com/1720
Reviewed-by: Eduardo Pinheiro <edpin@google.com>
Reviewed-by: Rob Pike <r@google.com>