path: make Path and String methods of Parsed do zero allocations
These methods are used everywhere but did two allocations on every
call (a buffer and a string).

Now, we compute the canonical representation once, and save it,
maintaining it in Drop and First.

We can almost always avoid the initial allocation to create the value,
too, because the parsed string is almost always clean or nearly so,
and we can use it, perhaps with a little trimming but no allocation.
Sometimes, though, allocation is required, but only once.

Because the cached value is not exported, we also provide an
UnmarshalJSON method for Parsed so it can be sent over the
wire and recreate the cached value.

More cleanups to come, but this CL achieves the biggest part.
All tests pass.

Change-Id: Ia2ee1e4f8264b9aa5470e14072b9a47d497c3ecb
Reviewed-on: https://upspin-review.git.corp.google.com/1733
Reviewed-by: Eduardo Pinheiro <edpin@google.com>