store/service: remember the user for each call
So we can tell who is calling each method, have Dial set up a different
service object identified by the calling user. This allows us to distinguish
who is calling when there are multiple users. Eventually this will be
authenticated.

The new information is unused here. This is a trial balloon for the
approach, which will have real consequences in the directory server
and its tests.

Change-Id: I171a2b69735a25afc9d5105aca91458a308935e1