cmd/dirserver: add authentication
Each network connection identifies the user for that connection,
and the directory server then can honor access controls for that user.
The authentication is honor system only; true authentication will come.
This CL just set ups a working mechanism.

The way this is done is to add an Authenticate method that must be
called first when establishing a new connection. That returns an ID
that is used to identify the service thread to process the request.

That last trick is done by having each ID invoke a different instance
of the Server object inside the RPC server, which is installed under
a name tagged with the corresponding ID.

Simple but effective.

Change-Id: I5293956b9fe6fa3fd18c070fdb216f25065bbebf
Reviewed-on: https://upspin-review.git.corp.google.com/1851
Reviewed-by: Eric Grosse <ehg@google.com>