upspin/proto: new package for protocol buffer definitions
At the moment, this is just a move from store/proto, but in
writing the user protos I found significant overlap and issues
that arose from that. It seems simpler to have a single place
for the pieces, and upspin/proto seems right since, to a high
approximation, these defintions are a mirror of those in the
upspin package itself.

Change-Id: I9bb2c4b7a7d694c8b5c36a7e424bfb40b22e9645