cmd/dfuse: Avoid extra open fds.  We really only need one per cached file.
Rather than keeping an open fd per open, we now keep one for anyt cached file
that is open independent of the number of opens. In the process I found
an fd that wan't being closed in some instances.

Also added some debugging.  Much of it will go once I'm happy that the fuse
implementation is working everywhere.

Change-Id: I61de806826f44c445b45fd9f40eb6314c856d653
Reviewed-on: https://upspin-review.git.corp.google.com/2021
Reviewed-by: Eduardo Pinheiro <edpin@google.com>
Reviewed-by: Rob Pike <r@google.com>