Update patch set 1
Patch Set 1:

Hi all.

This is the first thing I noticed looking at the codebase, so I thought I'd just send a change as a way to say hello.

To make this change work with "go get" we'd need to serve a custom <meta> tag in the HTML headers at upspin.io. I can arrange that if we decide to go ahead.

One issue: it's conceivable that we'd want some path component between the base "upspin.io"  and the rest of the package name. For instance, "upspin.io/foo/auth" might be preferable to just "upspin.io" in case "upspin.io/auth" becomes meaningful in some other capacity. I'm not sure what a good value of "foo" is, either. It could be "src", "code", or merely "x" (if we need it at all). 

adg

Patch-set: 1