Update patch set 3
Patch Set 3:

I think this works. I talked to Dave about the race between pinging and adding to the cache a stale version and he suggested a generation count and if they don't match, we retry. This might have consequences to the rest of the code when we have real concurrency because we may need to retry bind a few times. Or, we could just retry inside bind 3 times and give up.

Tested it with -race many times and instrumented and played with the random sleeps.

Patch-set: 3
Label: Code-Review=0