Update patch set 1
Patch Set 1: Code-Review+2

LG.

Unrelated request (for a new CL): we currently use letsencrypt for our certificates. They expire ever 3 or 4 months (the next expiration is coming up now in June). I have a long standing TODO to automate that. Russ wrote a module to auto-fetch certificates with some frequency. If we're going to use that, this is the place, since this server is what the letsencrypt servers would make their requests to.

If you're willing to do that, that would be great. I'll start a thread about it, this is not the place...

Thanks.

Patch-set: 1
Reviewer: Gerrit User 5015 <5015@0759ed4b-3fdc-34dc-866b-f2e708168c03>
Label: Code-Review=+2