Update patch set 2
Patch Set 2:

(5 comments)

> (5 comments)
 > 
 > isn't cmd/store/grpcstore/grpcstore.go the same as cmd/storeserver?
 > if not, what is the difference?

It's subtle. It has to do with that email I sent earlier today. Basically, we can't really use bind and implement a storeFor(gContext) here. I was going that route and hit a wall. Basically, it could be if we moved most of cmd/store into store/gcp or some other store/XXX in parallel with teststore.

Basically, when you register a new service type, say type upspin.GCP, you're registering something that implements the interface directly, not a client that makes networking calls. In order for this to work, I need to move code around as per option b) in that email.

It's way easier to explain this in person, looking at code. It might take 20 minutes, but it has to do with the hall of mirrors you talk about. Who implements which interface. 

Short story: yes, it will be at some point the same as storeserver. But it will require a slightly larger ROTD. This is a step closer though.

Patch-set: 2