all: introduce "transports" package for each of directory, store, and user
Shortens the code and makes the things that need many imports easier
to maintain.

It could and maybe should be used in a few more places; I installed them
only in places where there would be zero effective change in the imported
packages.

We could do the same for packers too, and might consider it.

Change-Id: I529e953dddaf3dc01dff4f6aaf7f3d1e19fd25b3
Reviewed-on: https://upspin-review.git.corp.google.com/2022
Reviewed-by: Eduardo Pinheiro <edpin@google.com>