cmd/uspin: add upspin share
This subcommand reports the users with read permission for each argument,
descending recursively into directories if requested.

If given the -fix flag, it will replace the wrapped keys in each file to match
the users defined by the Access files.

Note that we do not have the API yet to discover what keys are already there,
so -fix will always overwrite the directory entry. This will be easy to fix in time.

This subcommand serves as a diagnostic but also as a stopgap while we
await automatic update from clients.

Change-Id: I79dafdf818071d910496500dc26ab5b11c0cee1d