auth/grpcauth: add option keep alive to connections
DO NOT REVIEW: publishing this as an RFC.

This is an application-layer keep-alive, similar to Andrew's TCP
keep-alive (https://upspin-review.git.corp.google.com/#/c/2160/) but
suitable for platforms where TCP does not implement keep alive (I don't
know of any, but who knows if older cellphones or IoT toasters do).

We may want to have it around but turned off, as an option.

This is not ready for submission though.

Change-Id: I9e274faec370370bcc58df84f63173d320729333