pack/ee: fix Share
The upspin share command now works. The bug was that for a long list of users,
the rebuild in Share would complete but would not clear the "skipped" bit for
that instance of packdata, so the call would fail.

This needs a test that is beyond my ability at the moment, but I know how
to recreate the problem with 'upspin share'.

Also fix a simple bug in 'upspin share' and print the packing in 'upspin ls -l'.

Change-Id: I28e0a3518f1c9a18b9cbec483d0cfbc26a49b9e4
Reviewed-on: https://upspin-review.git.corp.google.com/2170
Reviewed-by: Eduardo Pinheiro <edpin@google.com>