Update patch set 1
Patch Set 1:

> (1 comment)

Oh, scratch all of that.

I think tcp_retries2 is a global setting that we can set on the kernel for all connections. It would be a hack for GCP/GCE, not something we want to put in our code.

I'll play around with this to see if I notice anything in my tests.

Patch-set: 1