store/gcp: fix lock copying bug
The lock is now global for the package. It could be a pointer lock that
gets copied but this is simpler. directory/gcp has a similar lock.

Note that a package lock means one address space cannot call bind.Register on
two Store services using different instances of GCP (say, one using one
bucket and another using another bucket). This is probably okay since GCP
is not an end-user transport anymore. It's a server transport. In other words,
users will use the Remote transport to talk to a GCP server.

If we want to relax this restriction later, we can make the lock and
refCount be a pointer for a given set of configuration instances.

Also: fixed the race condition in the test that happened for Rob once.
Ran 100's of times to verify.

All tests pass. No new lint or vet errors.

Change-Id: I344687fc9bec91373ffeec1cd863124eee24d85b