cmd/upspin: rewrite comment for unknown user in share
We do not plan to provide a way to look up a user by key, since it is
both difficult and unnecessary. What is important here is that if we
discover unrecognized users in the packing, we should report that
the file needs repacking so it can be fixed soon.

Change-Id: Ie1d3dd406eb9711c3ee5733ea87d4fd19b0cfe4f