errors: new package
This is a proposal for a new errors package to be used in all of Upspin.
The idea is that it carries information about the path, the user, the
operation attempted, and what sort of error it is (permission, I/O, etc.).
The constructer, E, makes it simple to fill in whatever fields make sense,
and also to cascade errors when needed.

For discussion.

Change-Id: I81099c7046e7d2be78541fbac12f00ec6f260a9c
Reviewed-on: https://upspin-review.git.corp.google.com/2260
Reviewed-by: Eduardo Pinheiro <edpin@google.com>
Reviewed-by: Andrew Gerrand <adg@google.com>