cmd/upspin: handle multiple keys better
This is tricky to test but works so far. The only assumption it's making is that
if we have a key, it's valid for one of the packings in the ee package.

Also add a -force flag to disregard the keys and recreate the mapping.
I suspect this will be handy to have one day.

Change-Id: Ib1febea4bf0219ff7c783b824c834d2e149fa5f3