errors: be careful about the string type
It would be easy for someone to have a user or path name with type
string rather than UserName or PathName. I don't know if it will
happen but it's easy to guard against: We notice if the Op has an @,
which should never happen, and fix it up, but also log the problem
so it can be found and fixed.

Change-Id: Ic4adb5bfc900fba51885145cd791623902e9ac8c