cmd/upspinfs: make access to files work even if we can't walk the path.
1) If we fail a Lookup due to permission, assume the element exists and is a directory.
This is hacky but I believe is the only way. It does lead to odd looking errors when
trying to write files that you have no permission to write, i.e., that the file is a
directory rather than permission denied.
2) Fix Rename, I was confusing the source and destination directory.  Found this when
trying to test the changes in 1.
3) Add a hack to avoid 0 length Access files by adding a new line when Put'ing them.
ALso found when trying to test 1.

Change-Id: I9cb029e717b7d633dadc31eaa15f14aa2a972539