errors: do not overwrite a valid path or user in the fallback string case
If the call is erroneous, we should not overwrite valid data.

We could panic here, but it seems safer to generate some error and log
so the proximate problem is visible now, and the long-term problem brought
to our attention. That is, this is an error we're trying to print, so get the error
printed.

Addressing post-submit comments from CL 2264.

Change-Id: Iec4bcfebc6514238d000d9965ac5b41ce1e46ae0
Reviewed-on: https://upspin-review.git.corp.google.com/2280
Reviewed-by: Eduardo Pinheiro <edpin@google.com>