cmd/upspinfs: renaming directories is disallowed (at least for now)
It is a royal pain since we have to copy the whole directory tree
in order to rename (each dir entry contains a signature with the
name of the file).  After finishing the copy, we then have to do
a reverse walk to remove all the old DirEntries.  Doesn't seem
worth it unless someone comes up with a compelling reason.

Change-Id: Ib8f070de1b3a280a07ab816a8e24625bb36d751b