all/remote: use new Error type
Have all the remotes use the new Error type. None of the servers
will cooperate yet, but everything still passes.

Now, the error returned by any RPC is always a network or grpc error.
The application error is always returned in the response's Error field.
This makes the code a little more intricate but will improve error
reporting once all the other pieces are in place.

Change-Id: I36c7c1b2d77ae9e84b27e69cfde6818dd6ae08f8
Reviewed-on: https://upspin-review.git.corp.google.com/2292
Reviewed-by: Eduardo Pinheiro <edpin@google.com>
Reviewed-by: Andrew Gerrand <adg@google.com>