Update patch set 1
Patch Set 1:

(2 comments)

The paucity is largely because of errors being generated by pieces underneath that will define them.

I want to see how it plays out. For instance, if path generates a syntax error, the wrapper might want to dig in and pull that Kind up to the top, and maybe delete it from the inner one. But we need experience first.

Patch-set: 1
Label: Code-Review=0