factotum: remove PrivateKey and KeyPair from upspin.go.
All use of private keys should be in factotum.  When factotum came
in, we were supposed to stop using upspin.KeyPair but use
continued to grow in test cases.  Ameliorate.

This is an initial step in a series of CLs that will allow readers
to use different elliptic curves, e.g. p256 and p386, on a single
packed file.  We want each user to only have one key at a time.

Change-Id: Ibdc101d6e5f8541d2af07f0683bee50f88a99039
Reviewed-on: https://upspin-review.git.corp.google.com/2316
Reviewed-by: Rob Pike <r@google.com>