Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (turns out command-enter sends the comments, oops)
> 
> I'm also running:
> 
> $ protoc --version
> libprotoc 3.0.0
> 
> Have you tried 'go get -u github.com/golang/protobuf/protoc-gen-go' ?

Yes. 

Are you on Linux or Mac?

How did you upgrade to 3.0.0? I tried that brew recipe and nothing worked. If I revert to what was before then, it's this:

$ protoc --version
libprotoc 2.6.1

No amount of go get -u or brew update brings me to 3.0.0 in a sane way. :-(

Patch-set: 2
Label: Code-Review=0