Update patch set 1
Patch Set 1:

I think this code is correct as written. A sync.Mutex field in a struct that is passed around by pointer is a common and good technique.

The actual fix is to avoid the struct copy on line 144 of all_test.go, which seems entirely unnecessary to me. (Change the embedded field in type client to a *AuthClientService.)

Patch-set: 1
Label: Code-Review=0