all: add a new "unassigned" transport
This is implemented by a set of User/Directory/Store servers
that error out on every call than can return an error.

Its purpose is to allow us to specify in an rc file that a particular
service cannot be used, avoiding potential surprises. For instance,
in rc.fileserver one can now write:

	directory=unassigned
	store=unassigned

because those services are implemented by it directly.

This is now the zero value for Transport, which required renumbering
the transports.

We can next use this in InitContext so contexts that do not define a
transport get this one.

Change-Id: Ia9a5a605b4cc8d3b31afd629c661b4f208d86e4e
Reviewed-on: https://upspin-review.git.corp.google.com/2462
Reviewed-by: David Presotto <p@google.com>
Reviewed-by: Eduardo Pinheiro <edpin@google.com>