pack/ee/ee.go: Add writer to packdata.
We have wrongly assumed that owner==writer.  (Here owner is the
prefix of pathname;  writer is the caller of Put.)  The fix is
easy, and will allow a better "ls" someday.

As agreed, this puts Writer in Metadata rather than Packdata.

All tests pass, but I'm not yet satisfied we have enough tests.

Change-Id: I8003bbfceff29ca64854cf64a939e80b5ecee366