pack/ee/ee.go: Add writer to packdata.
We have wrongly assumed that owner==writer.  (Here owner is the
prefix of pathname;  writer is the caller of Put.)  The fix is
easy, and will allow a better "ls" someday.

As agreed, this puts Writer in Metadata rather than Packdata.

Removed the unused (and no longer useful) PackingString method
from Factotum.

Left curveName for now in the struct, since ECDSA is also there.
Someday when we need to extend crypto, other changes will also
be needed, but that can wait.  I think this is now ready to submit.

(No doubt there will be some debugging yet to do and tests to add.)

Change-Id: I8003bbfceff29ca64854cf64a939e80b5ecee366
Reviewed-on: https://upspin-review.git.corp.google.com/2464
Reviewed-by: Eduardo Pinheiro <edpin@google.com>