pack/ee/ee.go: Add writer to packdata.
*** Incompatible packdata format change.
*** Coordinate submit with storage wipe.

We have wrongly assumed that owner==writer.  (Here owner is the
prefix of pathname;  writer is the caller of Put.)  The fix is
easy, and will allow a better "ls" someday.

I'll discuss at the meeting whether to preserve wrapped keys
for owner and writer during Share;  I think we should but Dave
and I are not certain we had agreement with everyone.

Change-Id: I8003bbfceff29ca64854cf64a939e80b5ecee366