Update patch set 1
Patch Set 1:

I'm really confused about this. I can't see why this is any better than just using flags directly. Sure, you're not importing "flag" directly into a package, but you're depending on a package that depends on flag. So, you've essentially added a middle layer for exactly the same purpose, just with extra code that now needs to be maintained.

The concern with flags is that people abuse it if we allow it in packages. But some usages are valid such as setting the log level for all binaries. I'd much prefer we set the rules on when it's okay to use flags than to give people another mechanism for doing essentially the same thing that can also be abused just the same.

But I might be missing some insight. Let's talk in person so you can explain your concerns with flag and why this is better.

Patch-set: 1