Update patch set 1
Patch Set 1:

There are a several problems with defining flags in packages. The main program is not in charge of their settings or management. Related flags end up in different packages with inconsistent names or values. If main needs to know the values of flags in packages, upstream API is required, complicating program setup. Also, package-level flags are initialized in unpredictable order, which can cause misbehavior.

Moreover, allowing packages the ability to modify flags exacerbates the lack of control from the top level. The main package is the controller; it should be in charge of all settings. The only place a package should be able to play with flags is in its tests.

The proposed packages addresses all these issues and more.

- It provides a central place for all flags that define global behavior. They will be defined consistently, used consistently, and available consistently.

- Flag values flow to packages, never from them.

- All commands will have the same definition and meaning for all flags.

- Only one package needs to be edited to provide flags to all commands.

- A command requires only one import ("flags") to connect to all flags it needs.

- It provides a simple uniform API to set up global flags.

- It provides a simple uniform API to query global flags.

- It delays flag initialization until main is initialized, forcing packages to initialize themselves properly through clear APIs rather than flag-controlled backchannels. (This isn't really a new feature; Go flags have this property already. Plus I do admit nothing stops a package from calling flags.Foo during initialization, but that will only ever get the default, so it doesn't worry me.)

- It provides read-only access. Flag setting can be done only from the command line. (I will provide a way for tests to set them as required, but insist they not be settable from non-test programs except through the command line.)

There's probably more I can't remember.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@0759ed4b-3fdc-34dc-866b-f2e708168c03>
Label: Code-Review=0