store/gcp: resume uploads when restarting
DO NOT REVIEW

Instead of creating a new dir for the local cache, use
the same one. At startup, if there are left-over files,
uploads them to the backend. This improves robustness
of the Store server.

Also, instead of launching one go routine for each upload,
keep a fixed number of go routines around for uploading. This
ensures under heavy load we won't be thrashing on I/O (networking
and file).

All tests pass. No new lint or vet errors.

Change-Id: I394ad028c7294e97d6ea5cfe2f0fbfb45c2212c4