Update patch set 5
Patch Set 5:

I don't think this is the correct strategy.

On Kubernetes, if a container fails it is nuked and a new one is created. The temporary directory disappears with the container. So this doesn't do anything to help us with reliability.

I wonder whether it's a good idea to use local disk to cache writes at all. Really, the store shouldn't ack any writes until they are committed to permanent storage.

Instead of using local disk, what if we just piped the uploads directly through to cloud storage? We could still use a local disk cache for reads.

Patch-set: 5