Update patch set 5
Patch Set 5:

My thinking here is that the temporary storage will soon become a PD. That doesn't go away when the pod is recreated.

Uploading to the cloud adds another 500ms *from the server*. Storing locally seems necessary to me based on a performance argument.

In light of this, PTAL?

Patch-set: 5