client/gobind: add Go bindings for the Client
- Adds a wrapped client and utility functions for configuring
and creating a Client for other languages to use, most notably
Java for Android.

Notes:
1) Just declaring a struct does not mean it will be
instantiatable in Java. We need to create the New funcs too.
2) Slices of any kind except byte cannot be returned, even
inside structs in the current implementation of gomobile.
3) We cannot export any upspin.* type since the upspin package
contains uint64, which Java can't handle.
4) This is WIP, only read-only methods of Client are exported
for now. And only Remote is supported (likely the only we'll ever
need for a mobile client, but perhaps we'll need others for
other types of uses that are not mobile). TBD

Tested with a Java mobile client for Android.

Change-Id: Ic89fff0107e285fdde248d3015ebd7951ce3554e